requirements.txt
django==1.11.3
django-model-utils==3.0.0
git+https://github.com/mgaitan/django-admin-row-actions#egg=django-admin-row-actions
git+https://github.com/mgaitan/django-sneak.git@#egg=django-sneak
git+https://github.com/mgaitan/django-admincommand.git@filefield#egg=django-admincommand
git+https://github.com/mgaitan/django-versatileimagefield@one-dimension-resize#egg=django-versatileimagefield
django-annoying==0.10.3
django-anymail==0.11.1
django-attachments==1.1.1
django-dbbackup==3.1.3.
django-fancy-cache==0.9.0
git+https://github.com/mgaitan/django-material#egg=django-fmaterial
django-extensions==1.8.1
django-formtools==2.0
html2text==2016.9.19
django-taggit==0.22.1
django-geojson==2.10.0
git+https://github.com/mgaitan/django-hijack.git#egg=django-hijack
django-leaflet==0.22.0
django-localflavor==1.5.2
djangorestframework==3.6.3
jsonfield==2.0.2
nameparser==0.5.3
phonenumbers==8.6.0
Pillow==4.2.1
pyexcel-xlsx==0.4.1
Fabric3==1.14.post1
easyimap==0.6.3
factory_boy==2.9.2
pytest-django==3.1.2
psycopg2_binary==2.7.7
sentry-sdk==0.7.6